# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Windows Update Metadata API
  version: "1.0.0"
  description: |
    Self-hosted Windows Update metadata platform. Provides build discovery,
    file metadata, build diffing, change feed, and CDN-proxied downloads.
    All /v1/ endpoints require mutual TLS (mTLS) client certificate authentication.
servers:
  - url: https://localhost:8443
    description: Local development

security:
  - mTLS: []

paths:
  /healthz:
    get:
      summary: Health check
      security: []
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"

  /readyz:
    get:
      summary: Readiness check (verifies database connectivity)
      security: []
      responses:
        "200":
          description: Service is ready
        "503":
          description: Service is unavailable

  /v1/builds:
    get:
      summary: List and search catalog builds
      parameters:
        - name: search
          in: query
          schema: { type: string }
          description: Substring filter on build title
        - name: arch
          in: query
          schema: { type: string, enum: [amd64, x86, arm64] }
        - name: ring
          in: query
          schema: { type: string, enum: [Dev, Beta, ReleasePreview, Retail] }
        - name: stable
          in: query
          schema: { type: boolean }
          description: Filter to stable (non-Insider, non-cumulative) builds
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: order
          in: query
          schema: { type: string, enum: [discovered_at, build_number, created_at] }
      responses:
        "200":
          description: List of builds
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Build"

  /v1/builds/{uuid}:
    get:
      summary: Get a single build by UUID
      parameters:
        - name: uuid
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Build found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        $ref: "#/components/schemas/Build"
        "404":
          $ref: "#/components/responses/NotFound"

  /v1/builds/{uuid}/files:
    get:
      summary: List files for a build
      parameters:
        - name: uuid
          in: path
          required: true
          schema: { type: string }
        - name: with_urls
          in: query
          schema: { type: boolean }
          description: When true, resolves live CDN download URLs (calls Microsoft)
        - name: revision
          in: query
          schema: { type: integer }
          description: Required when with_urls=true
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/FileResult"

  /v1/builds/{uuid}/files/{filename}/download:
    get:
      summary: Stream a file from Microsoft CDN (proxied)
      parameters:
        - name: uuid
          in: path
          required: true
          schema: { type: string }
        - name: filename
          in: path
          required: true
          schema: { type: string }
        - name: revision
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Binary file stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  /v1/updates/fetch:
    post:
      summary: Trigger a live Windows Update query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchRequest"
      responses:
        "200":
          description: Fetch result summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        $ref: "#/components/schemas/FetchResult"

  /v1/diff:
    get:
      summary: Compare two builds' file sets
      parameters:
        - name: base
          in: query
          required: true
          schema: { type: string }
        - name: target
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: BuildDiff result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        $ref: "#/components/schemas/BuildDiff"

  /v1/feed:
    get:
      summary: Paginated build change-feed
      parameters:
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: event_type
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Feed entries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/FeedEntry"

  /v1/feed/stream:
    get:
      summary: Server-Sent Events live feed
      responses:
        "200":
          description: SSE stream of BuildEvent objects
          content:
            text/event-stream:
              schema:
                type: string

components:
  securitySchemes:
    mTLS:
      type: mutualTLS

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"

  schemas:
    Envelope:
      type: object
      properties:
        data: {}
        meta:
          $ref: "#/components/schemas/Meta"

    Meta:
      type: object
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }

    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }

    StatusResponse:
      type: object
      properties:
        status: { type: string }

    Build:
      type: object
      properties:
        uuid: { type: string }
        revision: { type: integer }
        title: { type: string }
        build: { type: string }
        arch: { type: string }
        ring: { type: string }
        flight: { type: string }
        is_stable: { type: boolean }
        is_insider: { type: boolean }
        is_cumulative: { type: boolean }
        discovered_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }

    FileResult:
      type: object
      properties:
        uuid: { type: string }
        name: { type: string }
        sha1: { type: string }
        sha256: { type: string }
        size_bytes: { type: integer }
        file_type: { type: string }
        url: { type: string, description: "Live CDN URL (populated when with_urls=true)" }
        expires_at: { type: string }

    FetchRequest:
      type: object
      properties:
        arch: { type: string, enum: [amd64, x86, arm64] }
        ring: { type: string, enum: [Dev, Beta, ReleasePreview, Retail] }
        flight: { type: string }
        build: { type: string }
        sku: { type: integer }

    FetchResult:
      type: object
      properties:
        total: { type: integer }
        new_builds:
          type: array
          items:
            $ref: "#/components/schemas/Build"
        updated:
          type: array
          items:
            $ref: "#/components/schemas/Build"

    BuildDiff:
      type: object
      properties:
        base_uuid: { type: string }
        target_uuid: { type: string }
        base_build: { type: string }
        target_build: { type: string }
        added:
          type: array
          items:
            $ref: "#/components/schemas/FileDiffEntry"
        removed:
          type: array
          items:
            $ref: "#/components/schemas/FileDiffEntry"
        changed:
          type: array
          items:
            $ref: "#/components/schemas/FileDiffEntry"
        unchanged_count: { type: integer }
        generated_at: { type: string, format: date-time }

    FileDiffEntry:
      type: object
      properties:
        name: { type: string }
        change_type: { type: string, enum: [added, removed, changed] }
        base: { $ref: "#/components/schemas/FileResult" }
        target: { $ref: "#/components/schemas/FileResult" }

    FeedEntry:
      type: object
      properties:
        id: { type: integer }
        event_type: { type: string }
        build_uuid: { type: string }
        build_title: { type: string }
        build_number: { type: string }
        arch: { type: string }
        ring: { type: string }
        occurred_at: { type: string, format: date-time }